move config from main to common
