+++ src/main/factories/repositories.ts
@@ -1,7 +1,7 @@
 import RatingRepository from "@/data/repositories/rating-repository";
 import UserRepository from "@/data/repositories/user-repository";
 import { FirebaeDatabase } from "@/infra/database/firebase-database";
-import config from "@/main/config";
+import config from "@/common/config";
